File reorg and function names
